Eliminando DELETE de Nacionalidad

+++ src/Spotify.Core/Genero.cs
@@ -2,6 +2,6 @@ namespace Spotify.Core;
 
 public class Genero
 {
-    public uint idGenero {get; set;}
+    public byte idGenero {get; set;}
     public required string genero {get; set;}
 }
